+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- program update drift-version to v2.132.0 to use MM oracle price for perps [#70](https://github.com/drift-labs/jit-proxy/pull/70)
+
 ### Fixes
 
 ### Breaking
@@ -82,7 +84,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixes
 
 ### Breaking
-
 
 ## [0.12.0] - 2024-12-06